backend/requirements.txt
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai
pydantic
pymongo
python-dotenv
python-dateutil
requests
pydantic[email]
